메뉴 페이징 구현

+++ app/src/main/java/com/hampson/dabokadmin/data/dto/MenuDto.kt
@@ -3,6 +3,6 @@ package com.hampson.dabokadmin.data.dto
 data class MenuDto(
     val id: Long? = null,
     val name: String? = null,
-    val type: Int? = null,
+    val type: Long? = null,
     val ingredient: Int? = null
 )
